+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The following is a history of the changes made to this project.
 
+## January 28th, 2016
+
+* The files making up the portal website have been overhauled completely.
+* Added a web based administration area which can be used to modify portal settings.
+* A simple blog system has been added which is functional but beta quality at best at this time.
+* A template system for the portal has been added which will over time evolve into a more user friendly system.
+* There is now a messages/aircraft/second graph displayed on the performance graphs web page.
+* Users can now specify the distance of range rings displayed on the dump1090 map during installation.
+* Uses the package collectd-core instead of collectd to eliminate the installation of unneeded packages.
+* Added modified the dump1090-mutability HTML files to be used as a dump978 map to the portal.
+* Added the ability to specify the heywhatsthat.com range rings ranges.
+
 ## January 20th, 2016
 
 * Added the ability to build and execute dump978.